requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
av==14.1.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
livekit==0.20.0
livekit-agents==0.12.13
livekit-api==0.8.2
livekit-plugins-openai==0.10.19
livekit-plugins-silero==0.7.4
livekit-protocol==0.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
onnxruntime==1.20.1
openai==1.63.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyreadline3==3.5.4
python-dotenv==1.0.1
sniffio==1.3.1
sympy==1.13.3
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==0.24.0
Werkzeug==3.1.3
yarl==1.18.3
